requirements.txt
django==1.7.0
djangorestframework==3.1.1
djoser==0.3.0
psycopg2==2.6
jsonfield==1.0.3
netaddr==0.7.10
mako==0.8.0
networkx==1.7
configobj==4.7.1
pyYAML==3.11
django-rest-swagger==0.3.4
django-radius
